@@ -1,4 +1,4 @@
-const { configure } = require('enzyme');
-const Adapter = require('enzyme-adapter-react-16');
+import Enzyme from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
 
-configure({ adapter: new Adapter() });
+Enzyme.configure({ adapter: new Adapter() });
