@@ -1,27 +1,29 @@
-import React, { Component } from 'react'
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
+import './BodySection.css';
 
 class BodySection extends Component {
+  constructor(props) {
+    super(props);
+  }
+
   render() {
+    const { children, title } = this.props;
     return (
       <div className='bodySection'>
-        <h2>{this.props.title}</h2>
-        {this.props.children}
+        <h2>{title}</h2>
+        {children}
       </div>
     );
   }
 }
 
-BodySection. defaultProps = {
-	children: <React.Fragment />
+BodySection.defaultProps = {
+  title: '',
 };
 
 BodySection.propTypes = {
-  title: PropTypes.string.isRequired,
-  children: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.element
-  ])
+  title: PropTypes.string,
 };
 
 export default BodySection;
