@@ -1,28 +1,28 @@
-import React, { Component } from 'react'
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import BodySection from './BodySection';
 import './BodySectionWithMarginBottom.css';
-import PropTypes from 'prop-types';
 
 class BodySectionWithMarginBottom extends Component {
+  constructor(props) {
+    super(props);
+  }
+
   render() {
     return (
       <div className='bodySectionWithMargin'>
-        <BodySection {...this.props}/>
+        <BodySection {...this.props} />
       </div>
     );
   }
 }
 
-BodySectionWithMarginBottom. defaultProps = {
-	children: <React.Fragment />
+BodySectionWithMarginBottom.defaultProps = {
+  title: '',
 };
 
 BodySectionWithMarginBottom.propTypes = {
-  title: PropTypes.string.isRequired,
-  children: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.element
-  ])
+  title: PropTypes.string,
 };
 
 export default BodySectionWithMarginBottom;
