@@ -1,12 +1,25 @@
-import React from "react";
+import React from 'react';
 import { shallow } from 'enzyme';
 import BodySectionWithMarginBottom from './BodySectionWithMarginBottom';
 
-describe("Testing BodySectionWithMarginBottom Component",() => {
-  it('checking that shallowing the component should render correctly a BodySection component and that the props are passed correctly to the child component', () => {
-    let wrapper = shallow(<BodySectionWithMarginBottom title="test title" />);
-		expect(wrapper.find("BodySection").exists()).toBe(true);
-    wrapper = shallow(<BodySectionWithMarginBottom title="test title"><p>test children</p></BodySectionWithMarginBottom>)
-		expect(wrapper.find("BodySection").props().title).toBe('test title');
+describe('<BodySectionWithMarginBottom />', () => {
+  it('component and props', () => {
+    const wrapper = shallow(
+      <BodySectionWithMarginBottom title='test title'>
+        <p>test children node</p>
+      </BodySectionWithMarginBottom>
+    );
+    const div = wrapper.find('.bodySectionWithMargin').first();
+    const BodySection = wrapper.find('BodySection');
+    const internalBody = BodySection.dive();
+    const h2 = internalBody.find('h2');
+    const p = internalBody.find('p');
+    expect(div.exists()).toEqual(true);
+    expect(BodySection).toHaveLength(1);
+    expect(BodySection.props().title).toEqual('test title');
+    expect(h2).toHaveLength(1);
+    expect(h2.text()).toEqual('test title');
+    expect(p).toHaveLength(1);
+    expect(p.text()).toEqual('test children node');
   });
 });
