@@ -1,38 +1,47 @@
 import React from 'react';
-import PropType from 'prop-types';
-import CourseShape from './CourseShape';
-import CourseListRow from './CourseListRow';
 import './CourseList.css';
+import CourseListRow from './CourseListRow';
+import PropTypes from 'prop-types';
+import CourseShape from './CourseShape';
 
-function CourseList({ listCourses }) {
+const CourseList = ({ listCourses }) => {
   return (
-    <table id="CourseList" cellPadding="0" cellSpacing="0">
+    <table className='course-list'>
       <thead>
-        <CourseListRow isHeader={true} textFirstCell='Available courses' />
-        <CourseListRow isHeader={true} textFirstCell='Course name' textSecondCell="Credit" />
+        <CourseListRow textFirstCell='Available courses' isHeader={true} />
+        <CourseListRow
+          textFirstCell='Course name'
+          textSecondCell='Credit'
+          isHeader={true}
+        />
       </thead>
       <tbody>
-        {
-        listCourses.length == 0 ?
-          <CourseListRow isHeader={false} textFirstCell='No course available yet'/>
-        : null
-        }
-        {
-          listCourses.map((val, idx) => {
-            return <CourseListRow isHeader={false} textFirstCell={val.name} textSecondCell={val.credit} key={val.id}/>
-          })
-        }
+        {listCourses.length === 0 && (
+          <CourseListRow
+            textFirstCell='No course available yet'
+            isHeader={false}
+          />
+        )}
+
+        {listCourses.map((course) => (
+          <CourseListRow
+            key={course.id}
+            textFirstCell={course.name}
+            textSecondCell={course.credit}
+            isHeader={false}
+          />
+        ))}
       </tbody>
     </table>
   );
-}
-
-CourseList.defaultProps = {
-  listCourses: []
 };
 
-CourseList.propType = {
-  listCourses: PropType.arrayOf(CourseShape)
+CourseList.defaultProps = {
+  listCourses: [],
+};
+
+CourseList.propTypes = {
+  listCourses: PropTypes.arrayOf(CourseShape),
 };
 
 export default CourseList;
