@@ -1,50 +1,49 @@
-// i this i have a problem here
-import { shallow } from "enzyme";
-import React from "react";
-import CourseListRow from "./CourseListRow";
+import React from 'react';
+import { shallow } from 'enzyme';
+import CourseListRow from './CourseListRow';
 
-describe("<CourseListRow />", () => {
-  it("CourseListRow renders without crashing", () => {
-    const wrapper = shallow(<CourseListRow textFirstCell="test" />);
-    expect(wrapper.exists()).toEqual(true);
+describe('<CourseListRow />', () => {
+  it('renders without crashing', () => {
+    const wrapper = shallow(<CourseListRow textFirstCell='test' />);
+    expect(wrapper.exists());
   });
-  it("When isHeader is true renders one cell with colspan = 2 when textSecondCell does not exist", () => {
+  it('renders one cell', () => {
     const wrapper = shallow(
-      <CourseListRow isHeader={true} textFirstCell="test" />
+      <CourseListRow isHeader={true} textFirstCell='test' />
     );
     wrapper.update();
-    const item = wrapper.find("th");
+    const th = wrapper.find('th');
 
-    expect(item).toHaveLength(1);
-    expect(item.prop("colSpan")).toEqual("2");
+    expect(th).toHaveLength(1);
+    expect(th.prop('colSpan')).toEqual('2');
   });
-  it("When isHeader is true renders two cells when textSecondCell is present", () => {
+  it('renders two cells', () => {
     const wrapper = shallow(
       <CourseListRow
         isHeader={true}
-        textFirstCell="test"
-        textSecondCell="second"
+        textFirstCell='test'
+        textSecondCell='second'
       />
     );
     wrapper.update();
-    const item = wrapper.find("th");
+    const th = wrapper.find('th');
 
-    expect(item).toHaveLength(2);
-    expect(item.first().text()).toEqual("test");
-    expect(item.at(1).text()).toEqual("second");
+    expect(th).toHaveLength(2);
+    expect(th.first().text()).toEqual('test');
+    expect(th.at(1).text()).toEqual('second');
   });
-  it("When isHeader is false renders correctly two td elements within a tr element", () => {
+  it('renders two td', () => {
     const wrapper = shallow(
       <CourseListRow
         isHeader={false}
-        textFirstCell="test"
-        textSecondCell="second"
+        textFirstCell='test'
+        textSecondCell='second'
       />
     );
     wrapper.update();
-    const item = wrapper.find("tr");
+    const tr = wrapper.find('tr');
 
-    expect(item).toHaveLength(1);
-    expect(item.children("td")).toHaveLength(2);
+    expect(tr).toHaveLength(1);
+    expect(tr.children('td')).toHaveLength(2);
   });
 });
