@@ -1,8 +1,9 @@
 import PropTypes from 'prop-types';
 
-export default PropTypes.shape({
-  id: PropTypes.number.isRequired,
-  name: PropTypes.string.isRequired,
-  credit: PropTypes.number.isRequired
-}
-);
+const CourseShape = PropTypes.shape({
+  id: PropTypes.number,
+  name: PropTypes.string,
+  credit: PropTypes.number,
+});
+
+export default CourseShape;
