@@ -2,18 +2,17 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import Footer from './Footer';
 
-describe("Testing <Footer /> component", () => {
-  let wrapper;
+const wrapper = shallow(<Footer />);
 
-  beforeEach(() => {
-    wrapper = shallow(<Footer />);
-  });
+it('renders without crashing', () => {
+  shallow(<Footer />);
+});
 
-  it("Footer Component renders without crashing", () => {
-    expect(wrapper.exists());
-  });
+it('renders footer', () => {
+  expect(wrapper.find('footer.footer').exists()).toEqual(true);
+});
 
-  it("Footer compoenent render at the very least the text “Copyright”", () => {
-    expect(wrapper.find("Copyright").at(0)).toBeDefined();
-  });
+it('renders footer', () => {
+  expect(wrapper.find('footer.footer p').exists()).toEqual(true);
+  expect(wrapper.find('footer.footer p').text()).toContain('Copyright');
 });
