@@ -1,23 +1,24 @@
-import React, { Component } from 'react'
+import React from 'react';
 
-const WithLogging = (WrappedComponent) => {
-	const name = WrappedComponent.displayName || WrappedComponent.name || 'Component';
+function WithLogging(WrappedComponent) {
+  const componentName =
+    WrappedComponent.displayName || WrappedComponent.name || 'Component';
 
-	class NewComponent extends Component {
-		componentDidMount() {
-			console.log(`Component ${name} is mounted`);
-		}
+  class HOC extends React.Component {
+    componentDidMount() {
+      console.log(`Component ${componentName} is mounted`);
+    }
 
-		componentWillUnmount() {
-			console.log(`Component ${name} is going to unmount`);
-		}
+    componentWillUnmount() {
+      console.log(`Component ${componentName} is going to unmount`);
+    }
 
-		render() {
-			return <WrappedComponent {...this.props}/>;
-		}
-	};
-	NewComponent.displayName = `WithLogging(${name})`;
-	return NewComponent;
+    render() {
+      return <WrappedComponent {...this.props} />;
+    }
+  }
+  HOC.displayName = `WithLogging(${componentName})`;
+  return HOC;
 }
 
-export default WithLogging
+export default WithLogging;
