@@ -1,14 +1,14 @@
 import React from 'react';
-import logo from '../assets/holberton_logo.jpg';
 import './Header.css';
+import logo from '../assets/holberton-logo.jpg';
 
 function Header() {
   return (
-    <div className="App-header">
-      <img src={logo} alt="logo" />
+    <header className='header'>
+      <img src={logo} alt='logo' />
       <h1>School dashboard</h1>
-    </div>
-  )
+    </header>
+  );
 }
 
 export default Header;
