@@ -1,24 +1,21 @@
 import React from 'react';
-import Header from './Header';
 import { shallow } from 'enzyme';
+import Header from './Header';
 
+const wrapper = shallow(<Header />);
 
-describe('Testing <Header /> Component', () => {
-  let wrapper;
+it('renders without crashing', () => {
+  shallow(<Header />);
+});
 
-  beforeEach(() => {
-    wrapper = shallow(<Header />);
-  });
+it('renders header', () => {
+  expect(wrapper.find('header.header').exists()).toEqual(true);
+});
 
-  it("Renders with out crashing", () => {
-    expect(wrapper).toBeDefined();
-  });
+it('renders header', () => {
+  expect(wrapper.find('header.header h1').exists()).toEqual(true);
+});
 
-  it("Render an h1 tag", () => {
-    expect(wrapper.find('h1')).toBeDefined();
-  });
-
-  it("Render an img tag", () => {
-    expect(wrapper.find('img')).toBeDefined();
-  });
+it('renders header', () => {
+  expect(wrapper.find('header.header img').exists()).toEqual(true);
 });
