@@ -3,17 +3,15 @@ import './Login.css';
 
 function Login() {
   return (
-    <React.Fragment>
-      <div className="App-body">
-        <p>Login to access the full dashboard</p>
-        <label htmlFor="email">Email: </label>
-        <input type="email" id="email" name="email" />
-        <label htmlFor="password">Password: </label>
-        <input type="password" id="password" name="password" />
-        <button>OK</button>
-      </div>
-    </React.Fragment>
-  )
+    <main role='main' className='login'>
+      <p>Login to access the full dashboard</p>
+      <label htmlFor='email'>Email</label>
+      <input type='email' name='email' id='email' />
+      <label htmlFor='password'>Password</label>
+      <input type='password' name='password' id='password' />
+      <button type='button'>OK</button>
+    </main>
+  );
 }
 
 export default Login;
