@@ -5,36 +5,44 @@ class NotificationItem extends React.PureComponent {
   constructor(props) {
     super(props);
   }
-
   render() {
-    return (
-      this.props.value ? 
-      <li
-      data-notification-type={this.props.type}
-      onClick={() => this.props.markAsRead(this.props.id)}
-      >{this.props.value}</li> 
-      :
-      <li
-      data-notification-type={this.props.type}
-      dangerouslySetInnerHTML={this.props.html}
-      onClick={() => {console.log('empty func');}}
-      ></li>
-    );
+    const { type, html, value, markAsRead, id } = this.props;
+    let li;
+
+    value
+      ? (li = (
+          <li data-notification-type={type} onClick={() => markAsRead(id)}>
+            {value}
+          </li>
+        ))
+      : (li = (
+          <li
+            data-notification-type={type}
+            dangerouslySetInnerHTML={html}
+            onClick={() => markAsRead(id)}
+          ></li>
+        ));
+
+    return li;
   }
 }
 
 NotificationItem.defaultProps = {
   type: 'default',
-  markAsRead: () => {console.log('empty func');},
-	id: 0
+  html: {},
+  value: '',
+  markAsRead: () => {},
+  id: NaN,
 };
 
 NotificationItem.propTypes = {
-  html: PropTypes.shape({__html: PropTypes.string}),
-  type: PropTypes.string.isRequired,
+  type: PropTypes.string,
+  html: PropTypes.shape({
+    __html: PropTypes.string,
+  }),
   value: PropTypes.string,
   markAsRead: PropTypes.func,
-  id: PropTypes.number
+  id: PropTypes.number,
 };
 
 export default NotificationItem;
