@@ -2,22 +2,45 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import NotificationItem from './NotificationItem';
 
-describe("Testing <NotificationItem />", () => {
-  let  wrapper;
-
-  it("<NotificationItem /> renders without crashing", () => {
-    wrapper = shallow(<NotificationItem />);
-    expect(wrapper.exists());
+describe('<NotificationItem />', () => {
+  it('renders without crashing', () => {
+    const wrapper = shallow(<NotificationItem />);
+    shallow(<NotificationItem />);
   });
 
-  it("<NotificationItem />  renders the correct html by passing dummy type and value props,", () => {
-    wrapper = shallow(<NotificationItem type="default" value="test" />);
-    expect(wrapper.find("li").text()).toBe("test");
-    expect(wrapper.find("li").prop("data-notification-type")).toBe("default");
+  it('renders type and value props', () => {
+    const wrapper = shallow(<NotificationItem type='default' value='test' />);
+    const li = wrapper.find('li');
+    expect(li).toHaveLength(1);
+    expect(li.text()).toEqual('test');
+    expect(li.prop('data-notification-type')).toEqual('default');
   });
 
-  it("<NotificationItem />  renders the correct html by passing a dummy html prop,", () => {
-    wrapper = shallow(<NotificationItem html={{__html:"<u>test</u>"}} />);
-    expect(wrapper.find("li").html()).toBe("<li data-notification-type=\"default\"><u>test</u></li>");
+  it('renders html prop', () => {
+    const text = 'Here is the list of notifications';
+    const wrapper = shallow(
+      <NotificationItem html={{ __html: '<u>test</u>' }} />
+    );
+    const li = wrapper.find('li');
+    expect(li.html()).toEqual(
+      '<li data-notification-type="default"><u>test</u></li>'
+    );
   });
 });
+/*
+describe('markAsRead', () => {
+  it('markAsRead message', () => {
+    const id = 16;
+    const wrapper = shallow(
+      <NotificationItem type='default' value='test' id={id} />
+    );
+    const instance = wrapper.instance();
+    instance.markAsRead = jest.fn();
+    const listItem = wrapper.find('li').first();
+    listItem.simulate('click');
+    instance.markAsRead(id);
+    expect(instance.markAsRead).toHaveBeenCalledWith(16);
+    jest.restoreAllMocks();
+  });
+});
+*/
